--- packages/runner-ct/src/main.scss
@@ -1,1 +1,0 @@
-@import 'specs/specs-list.scss';
